third_party/WebKit/LayoutTests/security/immutable-prototype-worker.html

<!DOCTYPE html>
<html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
'use strict';

let expected = ['immutable', 'immutable', 'immutable', 'immutable', 'immutable'];

async_test(function(t) {
  let worker = new Worker('immutable-prototype-worker.js');
  worker.onmessage = t.step_func_done(function(result) {
    assert_array_equals(expected, result.data);
  });
}, 'worker prototype chain should be immutable');
</script>
</html>
